pyproject.toml
[project]
name = "fastapi-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.116.1",
    "google-generativeai>=0.8.5",
    "langchain-community>=0.3.27",
    "langchain-core>=0.3.74",
    "langchain-google-genai>=2.0.10",
    "langgraph>=0.6.5",
    "uvicorn>=0.35.0",
    "python-dotenv>=1.0.0",
    "ipython>=9.4.0",
]
